# inventory-service/inventory-infrastructure-rest/src/main/resources/static/openapi.yaml
openapi: 3.1.1
info:
  title: Swagger Inventory Service
  description: |-
    Bienvenido a la documentación oficial del microservicio 'Payment Service'. Aquí podrás explorar e interactuar con los 
    diversos recursos disponibles través de su API. 
    Esta documentación proporciona los detalles técnicos necesarios para utilizar las funcionalidades ofrecidas por el 
    servicio de manera eficiente.

    Algunos enlaces útiles:
      - [El repositorio de Inventory Service](https://github.com/ByLuisen/my-k8s-project/tree/main/apps/order-service)
      - [La definición de la API para Inventory Service](https://github.com/ByLuisen/my-k8s-project/blob/main/apps/order-service/src/main/resources/static/openapi.yaml)
  version: 1.0.12
servers:
  - url: /api
...
paths:
  /v1/suppliers/me:
    put:
      tags:
        - Suppliers
      summary: Actualiza tus datos de proveedor
      description: Actualiza tus datos de proveedor
      operationId: updateSupplier
      requestBody:
        description: Datos necesarios para realizar la actualización
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSupplierRequestDto'
        required: true
      responses:
        '200':
          description: Operación exitosa
          content:
            'applications/json':
              schema:
                $ref: '#/components/schemas/SupplierResponse'
        '401':
          description: No autenticado
        '403':
          description: El token de acceso no tiene los roles o permisos requeridos
        default:
          description: Error inesperado
      security:
        - BambiOAuth2: [supplier]
      x-roles-required:
        - SUPPLIER
    delete:
      tags:
        - Suppliers
      summary: Elimina tus datos de proveedor
      description: Elimina tus datos de proveedor
      operationId: deleteSupplierById
      responses:
        '204':
          description: Operación exitosa
        '401':
          description: No autenticado
        '403':
          description: El token de acceso no tiene los roles o permisos requeridos
        default:
          description: Error inesperado
      security:
        - BambiOAuth2: [supplier]
      x-roles-required:
        - SUPPLIER
  /v1/suppliers:
    get:
      tags:
        - Suppliers
      summary: Lista nuestros diferentes proveedores
      description: Lista nuestros diferentes proveedores
      operationId: getAllSuppliers
      parameters:
        - name: supplierRequestDto
          in: query
          description: Filtros para obtener nuestros proveedores
          required: true
          schema:
            $ref: '#/components/schemas/SupplierRequestDto'
      responses:
        '200':
          description: Operación exitosa
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/PaginatedResultResponseSupplierResponse'
        '400':
          description: Solicitud incorrecta
        '404':
          description: Proveedores no encontrados
        default:
          description: Error inesperado
  /v1/products/{productId}:
    get:
      tags:
        - Products
      summary: Obtén información acerca de un producto
      description: Obtén información acerca de un producto
      operationId: findProductWithDetails
      parameters:
        - name: productId
          in: path
          description: ID del producto
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Operación exitosa
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ProductWithDetailsReadResponse'
        '400':
          description: Solicitud incorrecta
        '404':
          description: Producto no encontrado
        default:
          description: Error inesperado
    put:
      tags:
        - Products
      summary: Actualiza la información de un producto
      description: Actualiza la información de un producto
      operationId: updateProduct
      parameters:
        - name: productId
          in: path
          description: ID del producto
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Datos necesarios para actualizar el producto
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductRequestDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Solicitud incorrecta
        '404':
          description: Producto no encontrado
        '401':
          description: No autenticado
        '403':
          description: El token de acceso no tiene los roles o permisos requeridos
        default:
          description: Error inesperado
      security:
        - BambiOAuth2: [admin]
      x-roles-required:
        - ADMIN
    delete:
      tags:
        - Products
      summary: Elimina un producto
      description: Elimina un producto
      operationId: deleteProduct
      parameters:
        - name: productId
          in: path
          description: ID del producto
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Operación exitosa
        '400':
          description: Solicitud incorrecta
        '404':
          description: Producto no encontrado
        '401':
          description: No autenticado
        '403':
          description: El token de acceso no tiene los roles o permisos requeridos
        default:
          description: Error inesperado
      security:
        - BambiOAuth2: [admin]
      x-roles-required:
        - ADMIN
  /v1/products:
    get:
      tags:
        - Products
      summary: Obtén la información de nuestros diferentes productos
      description: Obtén la información de nuestros diferentes productos
      operationId: findProducts
      parameters:
        - name: filters
          in: query
          description: Filtros disponibles
          required: true
          schema:
            $ref: '#/components/schemas/ProductRequestDto'
      responses:
        '200':
          description: Operación exitosa
          content:
            'application/json':
              schema:
                $ref: >-
                  #/components/schemas/PaginatedResultResponseProductSummaryReadResponse
        '400':
          description: Solicitud incorrecta
        '404':
          description: Productos no encontrados
        default:
          description: Error inesperado
    post:
      tags:
        - Products
      summary: Crea un nuevo producto
      description: Crea un nuevo producto
      operationId: saveProduct
      requestBody:
        description: Datos necesarios para crear un nuevo producto
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequestDto'
        required: true
      responses:
        '200':
          description: Operación exitosa
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          description: Solicitud incorrecta
        '401':
          description: No autenticado
        '403':
          description: El token de acceso no tiene los roles o permisos requeridos
        default:
          description: Error inesperado
      security:
        - BambiOAuth2: [admin]
      x-roles-required:
        - ADMIN
...
components:
  schemas:
    UpdateSupplierRequestDto:
      type: object
      properties:
        name:
          type: string
          maxLength: 50
          minLength: 3
          pattern: ^[A-Za-z0-9][A-Za-z0-9 .-]*[A-Za-z0-9]$
          examples: ['TechGear Partners']
      required:
        - name
    SupplierResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          examples: ['TechGear Partners']
    UpdateProductRequestDto:
      type: object
      properties:
        brandId:
          type: string
          format: uuid
          examples: ['27f2e42b-a68e-4c89-abe3-69ec78675fbb']
        departmentId:
          type: string
          format: uuid
          examples: ['fc7e4e78-46ac-47b6-b54f-b12a74a22aa7']
        productStatusId:
          type: string
          format: uuid
          examples: ['1e8b7e83-bf45-4838-80bd-aa9e2cb9c286']
        sku:
          type: string
          pattern: ^[A-Za-z0-9_-]+$
          examples: ['SAM-001-EL']
        name:
          type: string
          maxLength: 100
          minLength: 3
          pattern: ^[A-Za-z0-9 ]+$
          examples: ['Samsung Galaxy S23']
        description:
          type: string
          maxLength: 500
          minLength: 10
          examples: ['Último modelo de la serie Galaxy con pantalla AMOLED de 120Hz.']
      required:
        - brandId
        - departmentId
        - description
        - name
        - productStatusId
        - sku
    ProductResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          examples: ['839546df-e9ab-45c4-b5e0-06f10ca5c3d7']
        brandId:
          type: string
          format: uuid
          examples: ['27f2e42b-a68e-4c89-abe3-69ec78675fbb']
        departmentId:
          type: string
          format: uuid
          examples: ['fc7e4e78-46ac-47b6-b54f-b12a74a22aa7']
        productStatusId:
          type: string
          format: uuid
          examples: ['1e8b7e83-bf45-4838-80bd-aa9e2cb9c286']
        sku:
          type: string
          examples: ['SAM-001-EL']
        name:
          type: string
          examples: ['Samsung Galaxy S23']
        description:
          type: string
          examples: ['Último modelo de la serie Galaxy con pantalla AMOLED de 120Hz.']
...
    CreateProductRequestDto:
      type: object
      properties:
        brandId:
          type: string
          format: uuid
        departmentId:
          type: string
          format: uuid
        productStatusId:
          type: string
          format: uuid
        sku:
          type: string
          pattern: ^[A-Za-z0-9_-]+$
          examples: ['SAM-001-EL']
        name:
          type: string
          maxLength: 100
          minLength: 3
          pattern: ^[A-Za-z0-9 ]+$
          examples: ['Samsung Galaxy S23']
        description:
          type: string
          maxLength: 500
          minLength: 10
          examples: ['Último modelo de la serie Galaxy con pantalla AMOLED de 120Hz.']
      required:
        - brandId
        - departmentId
        - description
        - name
        - productStatusId
        - sku
...
    SupplierRequestDto:
      type: object
      properties:
        size:
          type: integer
          format: int32
          minimum: 1
          examples: [20]
        page:
          type: integer
          format: int32
          examples: [0]
    PaginatedResultResponseSupplierResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/SupplierResponse'
        count:
          type: integer
          format: int64
          examples: [1]
...
    ProductRequestDto:
      type: object
      properties:
        brandId:
          type: string
          pattern: >-
            ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12})(\|[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12})*$
          examples: ['d0c85be3-154b-4ce4-82ee-d7d1687ac198|de0531ee-576c-4686-8cd9-16dec1e71070']
        departmentId:
          type: string
          examples: ['f9c66d68-7047-470b-bbeb-f9862dfbd95b']
        productStatusId:
          type: string
          pattern: >-
            ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12})(\|[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12})*$
          examples: ['1e8b7e83-bf45-4838-80bd-aa9e2cb9c286']
        sku:
          type: string
          pattern: ^([a-zA-Z0-9-]{4,20})(\|[a-zA-Z0-9-]{4,20})*$
          examples: ['SAM-001-EL']
        orderBy:
          type: string
          pattern: (created_at|sku)+-(asc|desc)$
          examples: ['created_at-desc']
        page:
          type: integer
          format: int32
          examples: [0]
        size:
          type: integer
          format: int32
          examples: [20]
      required:
        - orderBy
        - page
        - size
    PaginatedResultResponseProductSummaryReadResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ProductSummaryReadResponse'
        count:
          type: integer
          format: int64
          examples: [1]
    ProductSummaryReadResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
          examples: ['SAM-001-EL']
        name:
          type: string
          examples: ['Samsung Galaxy S23']
        brand:
          type: string
          examples: ['Samsung']
        imageUrl:
          type: string
          examples: ['https://example.com/images/6e4573cc-06d3-48a7-a7b6-6e67f64f7045.jpg']
        createdAt:
          type: string
          format: date-time
          examples: ['2025-06-07T21:00:03.348+02:00']
    ProductWithDetailsBrandResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          examples: ['27f2e42b-a68e-4c89-abe3-69ec78675fbb']
        name:
          type: string
          examples: ['Samsung']
    ProductWithDetailsDepartmentResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          examples: ['fc7e4e78-46ac-47b6-b54f-b12a74a22aa7']
        name:
          type: string
          examples: ['Electronics']
    ProductWithDetailsReadResponse:
      type: object
      properties:
        sku:
          type: string
          examples: ['SAM-001-EL']
        name:
          type: string
          examples: ['Samsung Galaxy S23']
        description:
          type: string
          examples: ['Último modelo de la serie Galaxy con pantalla AMOLED de 120Hz.']
        brand:
          $ref: '#/components/schemas/ProductWithDetailsBrandResponse'
        department:
          $ref: '#/components/schemas/ProductWithDetailsDepartmentResponse'
        status:
          type: string
          examples: ['Active']
        images:
          type: array
          items:
            type: string
            examples: ['https://example.com/images/productos/camiseta-deportiva.jpg']
...
  securitySchemes:
    BambiOAuth2:
      type: oauth2
      flows:
        authorizationCode:
#          authorizationUrl: http://localhost:8081/realms/bambi/protocol/openid-connect/auth
#          tokenUrl: http://localhost:8081/realms/bambi/protocol/openid-connect/token
          authorizationUrl: http://keycloak.local/realms/bambi/protocol/openid-connect/auth
          tokenUrl: http://keycloak.local/realms/bambi/protocol/openid-connect/token
          scopes:
            - admin: Acceso administrativo
            - supplier: Acceso proveedor